test: extract shared Lexical CaptureEditorPlugin into prompt-editor test utils
