+++ web/app/components/base/prompt-editor/plugins/history-block/index.spec.tsx
@@ -1,15 +1,14 @@
 import type { LexicalEditor } from 'lexical'
 import type { RoleName } from './index'
 import { LexicalComposer } from '@lexical/react/LexicalComposer'
-import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext'
 import { act, render, waitFor } from '@testing-library/react'
 import {
   $getRoot,
   $nodesOfType,
 } from 'lexical'
-import { useEffect } from 'react'
 import { HISTORY_PLACEHOLDER_TEXT } from '../../constants'
 import { CustomTextNode } from '../custom-text/node'
+import { CaptureEditorPlugin } from '../test-utils'
 import {
   DELETE_HISTORY_BLOCK_COMMAND,
   HistoryBlock,
@@ -17,20 +16,6 @@ import {
   INSERT_HISTORY_BLOCK_COMMAND,
 
 } from './index'
-
-type CaptureEditorPluginProps = {
-  onReady: (editor: LexicalEditor) => void
-}
-
-const CaptureEditorPlugin = ({ onReady }: CaptureEditorPluginProps) => {
-  const [editor] = useLexicalComposerContext()
-
-  useEffect(() => {
-    onReady(editor)
-  }, [editor, onReady])
-
-  return null
-}
 
 const createRoleName = (overrides?: Partial<RoleName>): RoleName => ({
   user: 'user-role',
